backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr Zagrivnyy",email = "alexzag2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pytest-django (>=4.10.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "drf-spectacular (==0.27.1)",
    "requests (>=2.32.3,<3.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "boto3 (>=1.37.13,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-countries (>=7.6.1,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
